+++ src/main/java/EatPic/spring/domain/user/service/UserService.java
@@ -1,7 +1,7 @@
 package EatPic.spring.domain.user.service;
 
 import EatPic.spring.domain.user.User;
-import EatPic.spring.domain.user.dto.request.SignupRequest;
+import EatPic.spring.domain.user.dto.SignupRequestDTO;
 import EatPic.spring.domain.user.enums.UserStatus;
 import EatPic.spring.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +15,7 @@ public class UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
     
-    public User signup(SignupRequest request) {
+    public User signup(SignupRequestDTO request) {
         // 이메일 중복 검사
         if (userRepository.existsByEmail(request.getEmail())) {
             throw new IllegalArgumentException("이미 사용 중인 이메일입니다.");
